[Refactor] : debounce 기능 ViewModel으로 이동
Related to: #13
