+++ GitHubSearchApp/Presentation/ViewModelType.swift
@@ -1,0 +1,16 @@
+//
+//  ViewModelType.swift
+//  GitHubSearchApp
+//
+//  Created by chmini on 2022/10/24.
+//
+
+import Foundation
+import RxSwift
+
+protocol ViewModel {
+    associatedtype Input
+    associatedtype Output
+    
+    func transform(input: Input, disposeBag: DisposeBag) -> Output
+}
